fix clang-tidy errors

+++ test/coverage/communication/NotificationSinkTest.cpp
@@ -96,7 +96,7 @@ std::string get_random_string(size_t length) {
 		    "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 		    "abcdefghijklmnopqrstuvwxyz";
 		const size_t max_index = (sizeof(charset) - 1);
-		return charset[rand() % max_index];
+		return charset[static_cast<size_t>(rand()) % max_index];
 	};
 	std::string str(length, 0);
 	std::generate_n(str.begin(), length, randchar);
